# docs/openapi/schedule.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
...
  "paths" : {
...
    "/schedule/lessons/{id}/blockage" : {
      "post" : {
        "tags" : [ "Lessons" ],
        "summary" : "Заблокировать пару старостой — запретить геоотметку, посещаемость ставит только староста вручную (HEADMAN)",
        "operationId" : "blockLesson",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Пара заблокирована",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Урок не найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Нельзя заблокировать прошедшую или отменённую пару",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Lessons" ],
        "summary" : "Снять блокировку пары старостой (HEADMAN)",
        "operationId" : "unblockLesson",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Блокировка снята",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Урок не найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedule/lessons/{id}/restore" : {
      "patch" : {
        "tags" : [ "Lessons" ],
        "summary" : "Восстановить отменённый урок (HEADMAN/ADMIN)",
        "operationId" : "restoreLesson",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Урок восстановлен",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Урок не найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Недопустимый переход статуса",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/lessons/{id}/geo-block" : {
      "patch" : {
        "tags" : [ "Lessons" ],
        "summary" : "Включить/выключить гео-блокировку урока (HEADMAN/ADMIN)",
        "operationId" : "toggleGeoBlock",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GeoBlockRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Гео-блокировка изменена",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Урок не найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/lessons/{id}/cancel" : {
      "patch" : {
        "tags" : [ "Lessons" ],
        "summary" : "Отменить урок (HEADMAN/ADMIN). Допускается отмена PLANNED/ACTIVE/CLOSED.",
        "operationId" : "cancelLesson",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CancelLessonRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Урок отменён",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Урок не найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Урок уже отменён",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedule/groups/{groupId}/lessons" : {
      "get" : {
        "tags" : [ "Lessons" ],
        "summary" : "Получить расписание группы за диапазон дат (все роли)",
        "operationId" : "getLessons",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "dateFrom",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "dateTo",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "PLANNED", "ACTIVE", "CLOSED", "CANCELLED" ]
            }
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        }, {
          "name" : "assembler",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PagedResourcesAssemblerLessonResponse"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Расписание группы",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedModelEntityModelLessonResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Доступ запрещён",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Ресурс не найден",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "EntityModelLessonResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "scheduleItemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "subjectId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PLANNED", "ACTIVE", "CLOSED", "CANCELLED" ]
          },
          "dayOfWeek" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lessonNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startTime" : {
            "type" : "string"
          },
          "endTime" : {
            "type" : "string"
          },
          "weekType" : {
            "type" : "string",
            "enum" : [ "ALL", "ODD", "EVEN" ]
          },
          "room" : {
            "type" : "string"
          },
          "geoBlocked" : {
            "type" : "boolean"
          },
          "blockedByHeadman" : {
            "type" : "boolean"
          },
          "blockedByUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "blockedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "cancelReason" : {
            "type" : "string"
          },
          "cancelledBy" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cancelledAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "occurrenceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "assignmentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "semesterId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "assignedTeacherId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lessonType" : {
            "type" : "string"
          },
          "generation" : {
            "type" : "integer",
            "format" : "int64"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "current" : {
            "type" : "boolean"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "GeoBlockRequest" : {
        "type" : "object",
        "description" : "Запрос на переключение геоблокировки пары",
        "properties" : {
          "blocked" : {
            "type" : "boolean",
            "description" : "Флаг геоблокировки: true — отметка по геолокации заблокирована, false — разблокирована",
            "example" : true
          }
        },
        "required" : [ "blocked" ]
      },
      "CancelLessonRequest" : {
        "type" : "object",
        "description" : "Запрос на отмену одной пары",
        "properties" : {
          "reason" : {
            "type" : "string",
            "description" : "Причина отмены пары",
            "example" : "Преподаватель болен",
            "maxLength" : 512,
            "minLength" : 0
          }
        },
        "required" : [ "reason" ]
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "PageMetadata" : {
        "type" : "object",
        "properties" : {
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "PagedResourcesAssemblerLessonResponse" : {
        "type" : "object",
        "properties" : {
          "forceFirstAndLastRels" : {
            "type" : "boolean",
            "writeOnly" : true
          }
        }
      },
      "PagedModelEntityModelLessonResponse" : {
        "type" : "object",
        "properties" : {
          "_embedded" : {
            "type" : "object",
            "properties" : {
              "lessonResponseList" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/EntityModelLessonResponse"
                }
              }
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        }
      },
      "Link" : {
        "type" : "object",
        "properties" : {
          "href" : {
            "type" : "string"
          },
          "hreflang" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "deprecation" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "templated" : {
            "type" : "boolean"
          }
        }
      },
      "Links" : {
        "type" : "object",
        "additionalProperties" : {
          "$ref" : "#/components/schemas/Link"
        }
      },
      "ErrorResponse" : {
        "description" : "Стандартный формат ошибки API (RFC 9457 Problem Details)",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP статус код",
            "example" : 404
          },
          "type" : {
            "type" : "string",
            "description" : "URI типа ошибки",
            "example" : "https://api.rutcampustrack.ru/problems/resource-not-found"
          },
          "title" : {
            "type" : "string",
            "description" : "Краткое описание ошибки",
            "example" : "Ресурс не найден"
          },
          "detail" : {
            "type" : "string",
            "description" : "Детальное описание",
            "example" : "Группа с id=99 не найдена"
          },
          "instance" : {
            "type" : "string",
            "description" : "URI запроса, вызвавшего ошибку",
            "example" : "/api/academic/groups/99"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Время возникновения ошибки (UTC)",
            "example" : "2026-04-24T10:15:30Z"
          },
          "traceId" : {
            "type" : "string",
            "description" : "Correlation ID из MDC (для трассировки в логах)",
            "example" : "abc-trace-123"
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Ошибки валидации полей body DTO (только для 400)",
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            }
          },
          "field" : {
            "type" : "string",
            "description" : "Имя поля DTO, вызвавшего конфликт (только для 409, BUG-006-2)",
            "example" : "login"
          },
          "extras" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Дополнительные данные для клиента (счётчики каскадного удаления, retry-after, и т.п.)",
              "example" : {
                "scheduleItemsCount" : 3
              }
            },
            "description" : "Дополнительные данные для клиента (счётчики каскадного удаления, retry-after, и т.п.)",
            "example" : {
              "scheduleItemsCount" : 3
            }
          }
        }
      },
      "FieldError" : {
        "description" : "Ошибка валидации одного поля DTO (RFC 9457 Extension Member)",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "Имя поля DTO",
            "example" : "displayName"
          },
          "rejectedValue" : {
            "type" : "object",
            "description" : "Отклонённое значение"
          },
          "message" : {
            "type" : "string",
            "description" : "Локализованное сообщение об ошибке",
            "example" : "Имя не может быть пустым"
          }
        }
      }
    }
  }
}